requirements.txt
flask==2.3.3
python-dotenv==1.0.0
openai==1.3.0
chromadb==0.4.15
sentence-transformers==2.2.2
llama-index==0.9.15
pymupdf==1.23.8
python-docx==0.8.11
pdfplumber==0.10.0
nltk==3.8.1
tqdm==4.66.1
pandas==2.1.3
llama-index-embeddings-huggingface==0.1.3
llama-index-vector-stores-chroma==0.1.4
